fix: assertion and move remote store into internal
On-behalf-of: SAP <matthias.bruns@sap.com>
